RenderManager and namechange
Added the rendermanager and fixed a few names.
