+++ NeuralCreature/NeuralCreature/Debug/NeuralCreature.tlog/NeuralCreature.lastbuildstate
@@ -1,2 +1,2 @@
-#TargetFrameworkVersion=v4.0:PlatformToolSet=v141:EnableManagedIncrementalBuild=false:VCToolArchitecture=Native32Bit:WindowsTargetPlatformVersion=8.1
-Debug|Win32|C:\Users\jsors\Documents\Visual Studio 2017\Projects\NeuralCreature\|
+#TargetFrameworkVersion=v4.0:PlatformToolSet=v140:EnableManagedIncrementalBuild=false:VCToolArchitecture=Native32Bit:WindowsTargetPlatformVersion=8.1
+Debug|Win32|K:\GitHub Repositories\NeuralJetsonCreature\NeuralCreature\|
