+++ NeuralCreature/NeuralCreature/Main.cpp
@@ -1,0 +1,17 @@
+// NeuralCreature.cpp : Defines the entry point for the console application.
+//
+
+#include "stdafx.h"
+
+#include <btBulletDynamicsCommon.h>
+
+#include <iostream>
+#include "BulletTest.h"
+#include "NeuralCreature.h"
+
+int main(void) {
+
+	NeuralCreature world;
+	world.init();
+
+}
