--- NeuralCreature/NeuralCreature/OpenGlTest.cpp
@@ -1,248 +1,0 @@
-#include "stdafx.h"
-#include "OpenGLTest.h"
-#include <glm/glm.hpp>
-
-
-OpenGlTest::OpenGlTest()
-{
-}
-
-
-OpenGlTest::~OpenGlTest()
-{
-
-}
-
-void OpenGlTest::controls(GLFWwindow* window, int key, int scancode, int action, int mods)
-{
-	if (action == GLFW_PRESS)
-		if (key == GLFW_KEY_ESCAPE)
-			glfwSetWindowShouldClose(window, GL_TRUE);
-}
-
-GLFWwindow* OpenGlTest::initWindow(const int resX, const int resY)
-{
-	if (!glfwInit())
-	{
-		fprintf(stderr, "Failed to initialize GLFW\n");
-		return NULL;
-	}
-	glfwWindowHint(GLFW_SAMPLES, 4); // 4x antialiasing
-
-									 // Open a window and create its OpenGL context
-	GLFWwindow* window = glfwCreateWindow(resX, resY, "TEST", NULL, NULL);
-
-	if (window == NULL)
-	{
-		fprintf(stderr, "Failed to open GLFW window.\n");
-		glfwTerminate();
-		return NULL;
-	}
-
-	glfwMakeContextCurrent(window);
-	//glfwSetKeyCallback(window, controls);
-
-	// Get info of GPU and supported OpenGL version
-	printf("Renderer: %s\n", glGetString(GL_RENDERER));
-	printf("OpenGL version supported %s\n", glGetString(GL_VERSION));
-
-	glEnable(GL_DEPTH_TEST); // Depth Testing
-	glDepthFunc(GL_LEQUAL);
-	glDisable(GL_CULL_FACE);
-	glCullFace(GL_BACK);
-	return window;
-}
-
-void OpenGlTest::drawCube()
-{
-	int v = 2;
-	GLfloat vertices[] =
-	{
-		-v, -v, -v,   -v, -v,  v,   -v,  v,  v,   -v,  v, -v,
-		v, -v, -v,    v, -v,  v,    v,  v,  v,    v,  v, -v,
-		-v, -v, -v,   -v, -v,  v,    v, -v,  v,    v, -v, -v,
-		-v,  v, -v,   -v,  v,  v,    v,  v,  v,    v,  v, -v,
-		-v, -v, -v,   -v,  v, -v,    v,  v, -v,    v, -v, -v,
-		-v, -v,  v,   -v,  v,  v,    v,  v,  v,    v, -v,  v
-	};
-
-	GLfloat colors[] =
-	{
-		0, 0, 0,   0, 0, 1,   0, 1, 1,   0, 1, 0,
-		1, 0, 0,   1, 0, 1,   1, 1, 1,   1, 1, 0,
-		0, 0, 0,   0, 0, 1,   1, 0, 1,   1, 0, 0,
-		0, 1, 0,   0, 1, 1,   1, 1, 1,   1, 1, 0,
-		0, 0, 0,   0, 1, 0,   1, 1, 0,   1, 0, 0,
-		0, 0, 1,   0, 1, 1,   1, 1, 1,   1, 0, 1
-	};
-
-	static float alpha = 0;
-	//attempt to rotate cube
-	glRotatef(alpha, 0, 1, 0);
-
-	/* We have a color array and a vertex array */
-	glEnableClientState(GL_VERTEX_ARRAY);
-	glEnableClientState(GL_COLOR_ARRAY);
-	glVertexPointer(3, GL_FLOAT, 0, vertices);
-	glColorPointer(3, GL_FLOAT, 0, colors);
-
-	/* Send data : 24 vertices */
-	glDrawArrays(GL_QUADS, 0, 24);
-
-	/* Cleanup states */
-	glDisableClientState(GL_COLOR_ARRAY);
-	glDisableClientState(GL_VERTEX_ARRAY);
-	alpha += 1;
-}
-
-void OpenGlTest::display(GLFWwindow* window)
-{
-	while (!glfwWindowShouldClose(window))
-	{
-		// Scale to window size
-		GLint windowWidth, windowHeight;
-		glfwGetWindowSize(window, &windowWidth, &windowHeight);
-		glViewport(0, 0, windowWidth, windowHeight);
-
-		// Draw stuff
-		glClearColor(0.0, 0.8, 0.3, 1.0);
-		glClear(GL_COLOR_BUFFER_BIT | GL_DEPTH_BUFFER_BIT);
-
-		glMatrixMode(GL_PROJECTION_MATRIX);
-		glLoadIdentity();
-
-		gluPerspective(60, (double)windowWidth / (double)windowHeight, 0.1, 100);
-
-		glMatrixMode(GL_MODELVIEW_MATRIX);
-		glTranslatef(0, 0, -5);
-
-		drawCube();
-
-		// Update Screen
-		glfwSwapBuffers(window);
-
-		// Check for any input, or window movement
-		glfwPollEvents();
-	}
-}
-
-void OpenGlTest::start()
-{
-	GLFWwindow* window = initWindow(1024, 620);
-	if (NULL != window)
-	{
-		display(window);
-	}
-	glfwDestroyWindow(window);
-	glfwTerminate();
-	return;
-}
-
-void OpenGlTest::init()
-{
-	// start GL context and O/S window using the GLFW helper library
-	if (!glfwInit()) {
-		fprintf(stderr, "ERROR: could not start GLFW3\n");
-		return;
-	}
-
-	// uncomment these lines if on Apple OS X
-	/*glfwWindowHint(GLFW_CONTEXT_VERSION_MAJOR, 3);
-
-	glfwWindowHint(GLFW_CONTEXT_VERSION_MINOR, 2);
-	glfwWindowHint(GLFW_OPENGL_FORWARD_COMPAT, GL_TRUE);
-	glfwWindowHint(GLFW_OPENGL_PROFILE, GLFW_OPENGL_CORE_PROFILE);*/
-
-	GLFWwindow* window = glfwCreateWindow(640, 480, "Hello Triangle", NULL, NULL);
-	if (!window) {
-		fprintf(stderr, "ERROR: could not open window with GLFW3\n");
-		glfwTerminate();
-		return;
-	}
-	glfwMakeContextCurrent(window);
-
-	GLenum err = glewInit();
-	if (GLEW_OK != err)
-	{
-		/* Problem: glewInit failed, something is seriously wrong. */
-		fprintf(stderr, "Error: %s\n", glewGetErrorString(err));
-	}
-
-	fprintf(stdout, "Status: Using GLEW %s\n", glewGetString(GLEW_VERSION));
-
-	// start GLEW extension handler
-	glewExperimental = GL_TRUE;
-	glewInit();
-
-	// get version info
-	const GLubyte* renderer = glGetString(GL_RENDERER); // get renderer string
-	const GLubyte* version = glGetString(GL_VERSION); // version as a string
-	printf("Renderer: %s\n", renderer);
-	printf("OpenGL version supported %s\n", version);
-
-	// tell GL to only draw onto a pixel if the shape is closer to the viewer
-	glEnable(GL_DEPTH_TEST); // enable depth-testing
-	glDepthFunc(GL_LESS); // depth-testing interprets a smaller value as "closer"
-
-						  /* OTHER STUFF GOES HERE NEXT */
-	float points[] = {
-		0.0f,  0.5f,  0.0f,
-		0.5f, -0.5f,  0.0f,
-		-0.5f, -0.5f,  0.0f
-	};
-
-	GLuint vbo = 0;
-	glGenBuffers(1, &vbo);
-	glBindBuffer(GL_ARRAY_BUFFER, vbo);
-	glBufferData(GL_ARRAY_BUFFER, 9 * sizeof(float), points, GL_STATIC_DRAW);
-
-	GLuint vao = 0;
-	glGenVertexArrays(1, &vao);
-	glBindVertexArray(vao);
-	glEnableVertexAttribArray(0);
-	glBindBuffer(GL_ARRAY_BUFFER, vbo);
-	glVertexAttribPointer(0, 3, GL_FLOAT, GL_FALSE, 0, NULL);
-
-	const char* vertex_shader =
-		"#version 400\n"
-		"in vec3 vp;"
-		"void main() {"
-		"  gl_Position = vec4(vp, 1.0);"
-		"}";
-
-	const char* fragment_shader =
-		"#version 400\n"
-		"out vec4 frag_colour;"
-		"void main() {"
-		"  frag_colour = vec4(0.5, 0.0, 0.5, 1.0);"
-		"}";
-
-	GLuint vs = glCreateShader(GL_VERTEX_SHADER);
-	glShaderSource(vs, 1, &vertex_shader, NULL);
-	glCompileShader(vs);
-	GLuint fs = glCreateShader(GL_FRAGMENT_SHADER);
-	glShaderSource(fs, 1, &fragment_shader, NULL);
-	glCompileShader(fs);
-
-	GLuint shader_programme = glCreateProgram();
-	glAttachShader(shader_programme, fs);
-	glAttachShader(shader_programme, vs);
-	glLinkProgram(shader_programme);
-
-	while (!glfwWindowShouldClose(window)) {
-		// wipe the drawing surface clear
-		glClear(GL_COLOR_BUFFER_BIT | GL_DEPTH_BUFFER_BIT);
-		glUseProgram(shader_programme);
-		glBindVertexArray(vao);
-		// draw points 0-3 from the currently bound VAO with current in-use shader
-		glDrawArrays(GL_TRIANGLES, 0, 3);
-		// update other events like input handling 
-		glfwPollEvents();
-		// put the stuff we've been drawing onto the display
-		glfwSwapBuffers(window);
-	}
-	// close GL context and any other GLFW resources
-	glfwTerminate();
-	return;
-
-}
